+++ src/inspect_ai/_eval/evalset.py
@@ -105,6 +105,7 @@ def eval_set(
     log_shared: bool | int | None = None,
     bundle_dir: str | None = None,
     bundle_overwrite: bool = False,
+    log_dir_allow_dirty: bool | None = None,
     **kwargs: Unpack[GenerateConfigArgs],
 ) -> tuple[bool, list[EvalLog]]:
     r"""Evaluate a set of tasks.
@@ -194,6 +195,9 @@ def eval_set(
             by this eval set will be bundled into this directory.
         bundle_overwrite: Whether to overwrite files in the bundle_dir.
             (defaults to False).
+        log_dir_allow_dirty: If True, allow the log directory to contain
+            unrelated logs. If False, ensure that the log directory only contains logs
+            for tasks in this eval set (defaults to False).
         **kwargs: Model generation options.
 
     Returns:
@@ -286,6 +290,7 @@ def eval_set(
     retry_cleanup = retry_cleanup is not False
     max_connections = starting_max_connections(models, GenerateConfig(**kwargs))
     max_tasks = max_tasks if max_tasks is not None else max(len(models), 4)
+    log_dir_allow_dirty = log_dir_allow_dirty is True
 
     # prepare console/status
     console = rich.get_console()
@@ -344,7 +349,9 @@ def eval_set(
         # validate that:
         #  (1) All tasks have a unique identifier
         #  (2) All logs have identifiers that map to tasks
-        validate_eval_set_prerequisites(resolved_tasks, all_logs)
+        all_logs = validate_eval_set_prerequisites(
+            resolved_tasks, all_logs, log_dir_allow_dirty
+        )
 
         # see which tasks are yet to run (to complete successfully we need
         # a successful eval for every [task_file/]task_name/model combination)
@@ -413,7 +420,8 @@ def eval_set(
 
     # final sweep to remove failed log files
     if retry_cleanup:
-        cleanup_older_eval_logs(log_dir)
+        task_ids = {result.eval.task_id for result in results}
+        cleanup_older_eval_logs(log_dir, task_ids)
 
     # report final status
     success = all_evals_succeeded(results)
@@ -504,10 +512,13 @@ def list_latest_eval_logs(
 
 
 # cleanup logs that aren't the latest
-def cleanup_older_eval_logs(log_dir: str) -> None:
-    latest_completed_task_eval_logs(
-        logs=list_all_eval_logs(log_dir), cleanup_older=True
-    )
+def cleanup_older_eval_logs(log_dir: str, task_ids: set[str]) -> None:
+    logs = [
+        log
+        for log in list_all_eval_logs(log_dir)
+        if log.header.eval.task_id in task_ids
+    ]
+    latest_completed_task_eval_logs(logs=logs, cleanup_older=True)
 
 
 def latest_completed_task_eval_logs(
@@ -555,8 +566,10 @@ def latest_completed_task_eval_logs(
 #  (2) all log files have identifiers that map to tasks (so we know we
 #      are running in a log dir created for this eval_set)
 def validate_eval_set_prerequisites(
-    resolved_tasks: list[ResolvedTask], all_logs: list[Log]
-) -> None:
+    resolved_tasks: list[ResolvedTask],
+    all_logs: list[Log],
+    log_dir_allow_dirty: bool,
+) -> list[Log]:
     # do all resolved tasks have unique identfiers?
     task_identifiers: Set[str] = set()
     for task in resolved_tasks:
@@ -569,13 +582,18 @@ def validate_eval_set_prerequisites(
             task_identifiers.add(identifier)
 
     # do all logs in the log directory correspond to task identifiers?
-    for log in all_logs:
-        if log.task_identifier not in task_identifiers:
-            raise PrerequisiteError(
-                f"[bold]ERROR[/bold]: Existing log file '{basename(log.info.name)}' in log_dir is not "
-                + "associated with a task passed to eval_set (you must run eval_set "
-                + "in a fresh log directory)"
-            )
+    if log_dir_allow_dirty:
+        return [log for log in all_logs if log.task_identifier in task_identifiers]
+    else:
+        for log in all_logs:
+            if log.task_identifier not in task_identifiers:
+                raise PrerequisiteError(
+                    f"[bold]ERROR[/bold]: Existing log file '{basename(log.info.name)}' in log_dir is not "
+                    + "associated with a task passed to eval_set (you must run eval_set "
+                    + "in a fresh log directory). You can use the `--log-dir-allow-dirty` option to allow "
+                    + "logs from other eval sets to be present in the log directory."
+                )
+        return all_logs
 
 
 # yield a unique identifier for a task (used to pair resolved tasks to log files)